pyproject.toml
[tool.poetry]
name = "empowering-small-scale-kg"
version = "0.1.0"
description = ""
authors = ["albert.sawczyn <albert.sawczyn@pwr.edu.pl>"]

[tool.poetry.dependencies]
python = "3.10.*"
scikit-learn = "1.1.3"
pykeen = "1.10.1"
rich = "12.6.0"
networkx = {extras = ["default"], version = "2.8.8"}
nested-dict = "1.61"
seaborn = "0.12.2"
spacy = "3.4.4"
en_core_web_sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl"}
en_core_web_lg = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1-py3-none-any.whl"}
gensim = "4.3.0"
typer = {extras = ["all"], version = "0.7.0"}
ipykernel = "6.20.2"
ipywidgets = "8.0.4"
mpire = "2.6.0"
lightning-lite = "1.8.6"
wandb = "0.13.9"
libtmux = "0.21.1"
levenshtein = "0.21.0"
python-dotenv = "1.0.0"
fasttext-wheel = "0.9.2"
faiss-cpu = "1.7.4"
jupyter = "1.0.0"


[tool.poetry.dev-dependencies]
black = { version = "22.10.0", extras = ["jupyter"] }
isort = { version = "5.10.1", extras = ["pyproject"] }
mypy = "1.8.0"
poethepoet = "0.16.5"
pyflakes = "3.0.1"
pytest = "7.2.0"
coverage = {extras = ["toml"], version = "6.5.0"}
pytest-env = "0.8.1"
dvc = { version = "2.36.0", extras = ["s3"] }

[tool.poetry.group.dev.dependencies]
papermill = "2.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
black = "black --check ./"
isort = "isort .  -c"
black_fix = "black ./"
isort_fix = "isort . "
pyflakes = "pyflakes mgi"
mypy = "mypy --install-types --non-interactive --explicit-package-bases mgi experiments"
coverage_test = "coverage run -m pytest"
coverage_report = "coverage report -mi"
test = ["coverage_test", "coverage_report"]
check = ["black", "isort", "pyflakes", "mypy"]
fix = ["black_fix", "isort_fix"]
all = ["check", "test"]
fix_check = ["fix", "check"]

[tool.black]
line-length = 100

[tool.isort]
line_length = 100
known_first_party = "mpi"
known_third_party = ["click", "pytest"]
profile = "black"

[tool.mypy]
python_version = "3.10"
strict = true
plugins = "numpy.typing.mypy_plugin"

[[tool.mypy.overrides]]
module = [
    "networkx",
    "nested_dict",
    "sklearn.neighbors",
    "pandas",
    "srsly",
    "mpire",
    "lightning_lite",
    "fasttext",
    "faiss",
    "datasets",
]
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = [
    "pykeen.*"
]
implicit_reexport = true
